pyproject.toml
[project]
name = "documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.30",
    "langchain-google-genai>=4.2.2",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.2.2",
]
